move data management to options
