+++ src/options/options.html
@@ -11,6 +11,7 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
     <link rel="stylesheet" type="text/css" href="../shared/css/vars.css">
+    <link rel="stylesheet" type="text/css" href="../shared/css/loader.css">
     <link rel="stylesheet" type="text/css" href="options.css">
 </head>
 
@@ -80,7 +81,7 @@
                 <div class="col-6">
                     <textarea id="custom-title-textarea" style="font-family: 'Fira Code'; font-size: 0.8rem;"
                         class="w-100 mb-2" rows="4"></textarea>
-                    <div class="flex-center-evenly mb-3" style="width: 60%">
+                    <div class="flex-center-evenly mb-3">
                         <button id="custom-title-save">Save function</button>
                         <button id="custom-title-default">Back to default</button>
                     </div>
@@ -97,6 +98,93 @@
                 <div class="col-12 mt-3">Result: <code id="custom-title-result"></code></div>
 
             </div>
+
+        </div>
+
+        <div class="section">
+            <h2>Data Management</h2>
+
+            <p>You can export your full Memory as a <code>json</code> file or as a bibliography.</p>
+            <p>To perform advanced data manipulation (paper merges, batch tagging etc.) you can process the exported
+                Memory file
+                and then load it into the extension, overwriting your current data.</p>
+
+
+            <div style="display: flex; align-items: center;" class="row mb-4">
+                <div class="col-6">
+                    <p class="mb-0"> &gt; Export the Memory as a BibTex <strong>bibliography</strong>: </p>
+                </div>
+                <div class="col-2">
+                    <button class="ms-2 flex-center-between code-font" id="download-bibtex-json">
+                        .json&nbsp;
+                        <svg viewBox="0 0 24 24" style="height: 18px; width: 18px; color: black; stroke: black">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                            <line x1="4" y1="20" x2="20" y2="20" />
+                            <line x1="12" y1="14" x2="12" y2="4" />
+                            <line x1="12" y1="14" x2="16" y2="10" />
+                            <line x1="12" y1="14" x2="8" y2="10" />
+                        </svg>
+                    </button>
+                </div>
+                <div class="col-2">
+                    <button class="ms-2 flex-center-between code-font" id="download-bibtex-plain">
+                        .bib&nbsp;
+                        <svg viewBox="0 0 24 24" style="height: 18px; width: 18px; color: black; stroke: black">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                            <line x1="4" y1="20" x2="20" y2="20" />
+                            <line x1="12" y1="14" x2="12" y2="4" />
+                            <line x1="12" y1="14" x2="16" y2="10" />
+                            <line x1="12" y1="14" x2="8" y2="10" />
+                        </svg>
+                    </button>
+                </div>
+            </div>
+
+            <div style="align-items: center;" class="row mb-4">
+                <div class="col-6">
+                    <p class="mb-0"> &gt; Export your full Memory <strong>data</strong>: </p>
+                </div>
+                <div class="col-2">
+                    <button class="ms-2 flex-center-between code-font" id="download-arxivmemory">
+                        .json&nbsp;
+                        <svg viewBox="0 0 24 24" style="height: 18px; width: 18px; color: black; stroke: black">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                            <line x1="4" y1="20" x2="20" y2="20" />
+                            <line x1="12" y1="14" x2="12" y2="4" />
+                            <line x1="12" y1="14" x2="16" y2="10" />
+                            <line x1="12" y1="14" x2="8" y2="10" />
+                        </svg>
+                    </button>
+                </div>
+            </div>
+
+            <div id="overwrite-container" class="row">
+                <div class="col-6" style="display: flex; align-items: center;">
+                    <p class="mb-0"> &gt; Load a <code>json</code> memory file (this will <strong>overwrite</strong>
+                        your
+                        Memory): </p>
+                </div>
+                <div class="col-2">
+                    <label id="file-upload-label" for="overwrite-arxivmemory-input" class="ms-2">Select
+                        file</label><input type="file" id="overwrite-arxivmemory-input" style="display: none;"></input>
+                </div>
+                <div class="col-2">
+                    <button id="overwrite-arxivmemory-button" disabled class="ms-2 flex-center-between code-font">
+                        .json&nbsp;
+                        <svg viewBox="0 0 24 24" style="height: 18px; width: 18px; color: black; stroke: black">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                            <path d="M4 17v2a2 2 0 0 0 2 2h12a2 2 0 0 0 2 -2v-2" />
+                            <polyline points="7 9 12 4 17 9" />
+                            <line x1="12" y1="4" x2="12" y2="16" />
+                        </svg>
+                    </button>
+                </div>
+                <div class="col-6"></div>
+                <div class="col-6 mt-2"><span id="overwrite-file-name"></span></div>
+
+                <div style="display: none; margin: auto; margin-top: 1rem" id="overwriteFeedback" class="col-11"></div>
+            </div>
+
 
         </div>
