+++ src/popup/js/popup.js
@@ -99,12 +99,6 @@ const setStandardPopupClicks = () => {
     });
 
     addListener("memory-switch", "click", handleMemorySwitchClick);
-
-    addListener("download-arxivmemory", "click", handleDownloadMemoryClick);
-    addListener("download-bibtex-json", "click", handleDownloadBibtexJsonClick);
-    addListener("download-bibtex-plain", "click", handleDownloadBibtexPlainClick);
-    addListener("overwrite-arxivmemory-button", "click", handleOverwriteMemory);
-    addListener("overwrite-arxivmemory-input", "change", handleSelectOverwriteFile);
 };
 
 /**
